requirements.txt
discord.py>=2.3.0
google-generativeai>=0.3.0
edge-tts>=6.1.0
aiohttp>=3.8.0
pydub>=0.25.0
